changed path length statistics counters to use 64-bit unsigned integers as they would otherwise overflow when tracing large numbers of paths.

+++ src/appleseed/renderer/kernel/lighting/drt/drtlightingengine.cpp
@@ -391,7 +391,7 @@ namespace
         const LightSampler&     m_light_sampler;
 
         uint64                  m_path_count;
-        Population<size_t>      m_path_length;
+        Population<uint64>      m_path_length;
     };
 }
